Feat/connect reactions (#297)
* feat(connect): reactions in 1:1 calls (shared module with meeting)

Extract the in-call reactions logic (send/float/broadcast + the "…" emoji
picker) into builtins/webrtc/reactions.js and mix it into both the meeting and
the connect windows via Object.assign. The meeting's inline copies are removed
(behavior unchanged — same methods, now on the prototype).

Connect wiring: reactions-picker wrapper + reaction-stack in the skeleton,
onUiEvent "react"/"reactions-more", onWsMessage "REACTION" → _applyRemoteReaction,
and _closeReactionsPicker on teardown. The reactions button and skin already
come from the shared meeting-shell topbar/mixin.


* fix(webrtc): team-call device change fails silently without mic permission

In a team call, changing the microphone or speaker and clicking Save did
nothing, with no error. Root cause: enumerateDevices() only fills in real
deviceId + label once microphone permission is granted; before that every row
in the picker carries an EMPTY deviceId. Picking a device then stores "" into
selectedInputDevice/selectedOutputDevice, so confirmDeviceSelection's
`if (this.selectedInputDevice)` / `if (this.selectedOutputDevice)` guards both
skip — the change is a silent no-op on BOTH input and output.

- Add ensureMediaPermission(): query the 'microphone' permission first (a fast
  no-op when already granted, so the normal path is unchanged) and only fall
  back to a short-lived getUserMedia probe (which triggers the prompt and
  refreshes the device ids/labels) when it isn't granted yet.
- updateAudioDevicesList(): enumerate only AFTER permission is granted, so the
  list carries real deviceIds; show a clear note when denied.
- confirmDeviceSelection(): verify permission before applying, and surface
  failures with Wm.alert instead of swallowing them — a failed change is no
  longer mistaken for "nothing happened".


* fix(webrtc): device-picker Confirm button invisible (transparent on white popup)

The Confirm button in the in-call mic/speaker picker used
`background-color: var(--btn-commit-background)`, but that token is malformed
in skin/vars/default.scss (`-primary-500` instead of `var(--primary-500)`), so
it computes to transparent (verified live: rgba(0,0,0,0)). White text on the
white popup made the button blank ("trùng màu nền").

Scope-safe fix: point the device-picker Confirm button at the intended commit
accent directly (var(--primary-500), with var(--primary-600) on hover) instead
of the broken token. NOTE: the malformed --btn-commit-background token also
transparents ~20 other commit buttons app-wide — left untouched here to avoid a
broad unverifiable recolor; worth a separate global fix.


* feat(tasks/meeting): list-view filter, column-watch bells, gantt start_date, mobile call layout

Lands the current integration working tree on test:

- Tasks · List filter (Figma 2099-50501): accordion filter popup with
  keyword search + priority/status (multi-select, OR within), due-date
  buckets (overdue/today/this week/this month) and files, plus the shared
  Assignee/member dimension. Applies on the List view; dimensions AND, values
  OR. New state _filters/_filterExpanded + filter-cat/filter-set/
  filter-keyword/filter-clear handlers.
- Tasks · Column-watch bells (Figma 2041-20161): a per-column header bell to
  subscribe to change-notifications for that column (server-backed, folder
  scoped). The activity panel renders task_column_change events ("updated a
  task you follow: <title>") and live-refreshes on the task.column_change WS
  push. New state _columnWatches + _loadColumnWatches/_toggleColumnWatch.
- Tasks · Gantt bars now span start_date→due (fallback to ctime when no
  start_date), matching the schema's Duration feature.
- Meeting · Mobile call layout: on phones the side panel starts closed so the
  call stage shows first; a live standalone call fills the viewport
  (position:fixed under the 48px topbar) and chat overlays the stage instead
  of shrinking it. Desktop unchanged.
- Locale: NOTIFY_COLUMN, TASK_COLUMN_CHANGE_ACTION across all six languages.


* fix(trash): empty-bin layout flicker — reload list in place instead of full re-feed

Deleting several files from Trash blinked the whole panel repeatedly, most
visibly the empty-state placeholder flickering out and back. Root cause
(confirmed live): every trash/restore/purge WebSocket echo runs _wsRefresh,
which did a full feed(require('./skeleton')) — tearing down and rebuilding the
entire panel scaffold (topbar + list + footer) on each echo. A multi-file
delete emits echoes that trickle in more than one debounce window apart, so
5 echoes → 5 full rebuilds → the layout blinked 5 times.

- _wsRefresh now reloads only the LIST part in place (list.restart()) and
  refreshes the count once it settles; the panel frame keeps its DOM nodes
  (verified: topbar/footer are the same nodes before/after), so the layout no
  longer blinks. Falls back to a full feed only when the list isn't mounted yet.
- Debounce 400 -> 600ms so echoes that land ~half a second apart coalesce into
  a single reload.


* fix(upload): progress popup X / Close / Cancel-all never dismiss in bundle mode

The upload-progress popup's close controls didn't close the window for a
drag-drop (bundle) upload. All three route to cancelAll(), and the bundle
branch cancelled the jobs then `return`ed WITHOUT calling goodbye() — only the
non-bundle path closed (delayed goodbye). Compounding it, the footer button is
relabeled "Close" on completion by mutating the DOM only; onUiEvent resolves
the service from the widget MODEL before the DOM attribute, so the relabeled
"Close" still fired the stale "cancel-all" service.

- cancelAll(): the bundle-mode branch now dismisses the popup (delayed
  goodbye) after cancelling, matching the non-bundle path — so the X (always
  service "cancel-all"), footer "Close" and "Cancel all" all close.
- Sync the widget model service ('close'/'cancel-all') wherever the footer
  button is relabeled (_refreshFooter + the bundle-complete relabel), so
  onUiEvent reads the correct action.


* feat(connect): screen-share in 1:1 calls + call-window UX fixes

Port the meeting's screen-share to the connect (P2P) window via a shared
module, and fix two call-window interaction bugs.

- webrtc/screenshare.js: extract the meeting's screen-share methods
  (startPresentation/stopPresentation/prepareRemoteScreen/onRemoteScreen*/
  removePresenter/_dockParticipants/_setShareLocked/_toggleScreenShareFullscreen/
  _stopOwnPresentation) into a shared module used by both windows. `super`
  calls become __room.prototype.X.call(this) (both windows extend the same
  base); meeting-only hooks (_setMemberPresenting/_updateFloatFocus/
  _clearFloatFocus/_uidForParticipant) are called through optional guards so
  connect skips them. Meeting behavior is unchanged.
- connect: mix in the module; onUiEvent handles start/stop-screenshare
  (one-at-a-time lock guard) and togglefullscreen.
- connect: override fitScreenSize/change_size (like the meeting) so sharing
  no longer auto-expands the window to ~full viewport — CSS owns the
  presenter layout, window geometry is left alone.
- chat-p2p: _onOutsideClick no longer dismisses the panel on clicks inside
  an active call window (.window-connect / .window-meeting); the call is
  usually launched from this panel and they coexist.
- webrtc skin: border-radius:0 on private-area room.


* small header call p2p window

---------
