+++ src/drumee/builtins/webrtc/skin/index.scss
@@ -124,6 +124,7 @@
 
     &[data-area="private"] {
       background-color: var(--normal-bg-80);
+      border-radius: 0;
     }
 
     &[data-area="dmz"] {
